# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion de boutique",
        "description": "Documentation de l'API REST pour la gestion de boutique (catégories, produits, acheteurs, achats). CCP 2026.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Créer un compte (rôle employé ou gestionnaire uniquement)",
                "operationId": "5d4fa5d5607ed675ba93907e1fc94534",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Amadou Ba"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "amadou.ba@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "employe",
                                        "enum": [
                                            "employe",
                                            "gestionnaire"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé, retourne l'utilisateur et son token"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Se connecter et obtenir un token",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@boutique.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie, retourne l'utilisateur et son token"
                    },
                    "422": {
                        "description": "Identifiants incorrects"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Se déconnecter (invalide le token actuel)",
                "operationId": "28d48c13fca984400ac8de99f54ee13a",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Récupérer les informations de l'utilisateur connecté",
                "operationId": "bc2032a379aa97707ccc61511897c120",
                "responses": {
                    "200": {
                        "description": "Informations de l'utilisateur"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/categories": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Lister toutes les catégories",
                "operationId": "fd346ca2a3db92af72505c9cf327969b",
                "responses": {
                    "200": {
                        "description": "Liste des catégories"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Créer une nouvelle catégorie",
                "operationId": "4ce53e714eb24fbdd4acb9706dbe7864",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Chaussures"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Toutes les chaussures de sport"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catégorie créée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès non autorisé (réservé gestionnaire/admin)"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Afficher le détail d'une catégorie (avec ses produits)",
                "operationId": "b6acb3f1c6a43306ab6fb1c408542a8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de la catégorie"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Catégorie introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Modifier une catégorie",
                "operationId": "a9374f00e39c11b554a16e4f1c0c6a94",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Chaussures de sport"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description mise à jour"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catégorie modifiée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "404": {
                        "description": "Catégorie introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Supprimer une catégorie",
                "operationId": "973cbeed53e164331725f7ceec9b7287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Catégorie supprimée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "404": {
                        "description": "Catégorie introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez le token obtenu via /login, sans le mot 'Bearer' devant (Swagger l'ajoute automatiquement).",
                "bearerFormat": "Sanctum token",
                "scheme": "bearer"
            }
        }
    },
...
}